+++ scripts/check-ip.php
@@ -3,17 +3,11 @@
 require_once __DIR__ . '/../vendor/autoload.php';
 
 use CrowdSecBouncer\Bouncer;
-use Symfony\Component\Cache\Adapter\PhpFilesAdapter;
-use Symfony\Component\Cache\Adapter\TagAwareAdapter;
 use Monolog\Formatter\LineFormatter;
 use Monolog\Handler\RotatingFileHandler;
 use Monolog\Handler\StreamHandler;
 use Monolog\Logger;
 
-
-// Init cache adapter
-
-$cacheAdapter = new TagAwareAdapter(new PhpFilesAdapter('', 0, __DIR__ . '/.cache'));
 
 // Parse argument
 
@@ -35,13 +29,12 @@ $fileHandler = new RotatingFileHandler(__DIR__ . '/crowdsec.log', 0, Logger::WAR
 $logger->pushHandler($fileHandler);
 
 // Init
-$bouncer = new Bouncer($cacheAdapter, $logger);
-
-$config = [
+$configs = [
     'api_key' => $bouncerKey,
     'api_url' => 'http://crowdsec:8080',
+    'fs_cache_path' => __DIR__ . '/.cache',
 ];
-$bouncer->configure($config);
+$bouncer = new Bouncer($configs, $logger);
 
 // Ask remediation to LAPI