+++ scripts/refresh-cache.php
@@ -5,19 +5,9 @@ require_once __DIR__ . '/../vendor/autoload.php';
 use CrowdSecBouncer\Bouncer;
 use Monolog\Handler\RotatingFileHandler;
 use Monolog\Logger;
-use Symfony\Component\Cache\Adapter\RedisTagAwareAdapter;
-use Symfony\Component\Cache\Adapter\TagAwareAdapter;
 
 // Configure paths
-$logPath = __DIR__.'/.crowdsec.log';
-$cachePath = __DIR__ . '/.cache';
-
-// Instantiate the "PhpFilesAdapter" cache adapter
-$cacheAdapter = new TagAwareAdapter(new Symfony\Component\Cache\Adapter\PhpFilesAdapter('', 0, $cachePath));
-// Or Redis: $cacheAdapter = new RedisTagAwareAdapter(RedisAdapter::createConnection('redis://your-redis-host:6379'));
-// Or Memcached: $cacheAdapter = new TagAwareAdapter(new MemcachedAdapter(MemcachedAdapter::createConnection
-//('memcached://your-memcached-host:11211')));
-// Parse argument
+$logPath = __DIR__ . '/.crowdsec.log';
 
 $bouncerKey = $argv[1];
 if (!$bouncerKey) {
@@ -26,16 +16,16 @@ if (!$bouncerKey) {
 
 // Instantiate the Stream logger with info level(optional)
 $logger = new Logger('example');
-$fileHandler = new RotatingFileHandler(__DIR__.'/crowdsec.log', 0, Logger::WARNING);
+$fileHandler = new RotatingFileHandler(__DIR__ . '/crowdsec.log', 0, Logger::WARNING);
 $logger->pushHandler($fileHandler);
 
 // Instantiate the bouncer
-$bouncer = new Bouncer($cacheAdapter, $logger);
-$bouncer->configure([
-        'api_key' => $bouncerKey,
-        'api_url' => 'http://crowdsec:8080'
-    ]
-);
+$configs = [
+    'api_key' => $bouncerKey,
+    'api_url' => 'http://crowdsec:8080',
+    'fs_cache_path' => __DIR__ . '/.cache'
+];
+$bouncer = new Bouncer($configs, $logger);
 
 // Refresh the blocklist cache
 $bouncer->refreshBlocklistCache();