+++ src/ApiClient.php
@@ -23,32 +23,42 @@ class ApiClient
 {
     /** @var LoggerInterface */
     private $logger;
+    /** @var array */
+    private $configs;
 
     /**
      * @var ClientAbstract
      */
     private $restClient;
 
-    public function __construct(LoggerInterface $logger, array $settings = [])
+    /**
+     * @param array $configs
+     * @param LoggerInterface $logger
+     * @throws BouncerException
+     */
+    public function __construct(array $configs, LoggerInterface $logger)
     {
         $this->logger = $logger;
-        $useCurl = !empty($settings['use_curl']);
-        $this->restClient = $useCurl ? new Curl($this->logger) : new FileGetContents($this->logger);
-    }
-
-    /**
-     * Configure this instance.
-     */
-    public function configure(string $baseUri, int $timeout, string $userAgent, string $apiKey): void
-    {
-        $this->restClient->configure($baseUri, [
-            'User-Agent' => $userAgent,
-            'X-Api-Key' => $apiKey,
+        $this->configs = $configs;
+        $useCurl = !empty($this->configs['use_curl']);
+        if (empty($this->configs['api_user_agent'])) {
+            throw new BouncerException('User agent is required');
+        }
+        $userAgent = $this->configs['api_user_agent'];
+        $this->configs['headers'] = [
+            'User-Agent' => $this->configs['api_user_agent'],
+            'X-Api-Key' => $this->configs['api_key'],
             'Accept' => 'application/json',
-        ], $timeout);
+        ];
+
+        $this->restClient = $useCurl ?
+            new Curl($this->configs, $this->logger) :
+            new FileGetContents($this->configs, $this->logger);
+
         $this->logger->debug('', [
             'type' => 'API_CLIENT_INIT',
             'user_agent' => $userAgent,
+            'rest_client' => \get_class($this->restClient)
         ]);
     }
 
@@ -67,13 +77,19 @@ class ApiClient
      * @throws BouncerException
      */
     public function getStreamedDecisions(
-        bool $startup = false,
+        bool  $startup = false,
         array $scopes = [Constants::SCOPE_IP, Constants::SCOPE_RANGE]
-    ): array {
+    ): array
+    {
         /** @var array */
         return $this->restClient->request(
             '/v1/decisions/stream',
             ['startup' => $startup ? 'true' : 'false', 'scopes' => implode(',', $scopes)]
         );
+    }
+
+    public function getRestClient(): ClientAbstract
+    {
+        return $this->restClient;
     }
 }