+++ src/StandaloneBounce.php
@@ -2,21 +2,13 @@
 
 namespace CrowdSecBouncer;
 
-use CrowdSecBouncer\Fixes\Memcached\TagAwareAdapter as MemcachedTagAwareAdapter;
 use ErrorException;
 use Exception;
 use IPLib\Factory;
-use Symfony\Component\Cache\Adapter\MemcachedAdapter;
-use Symfony\Component\Cache\Adapter\PhpFilesAdapter;
-use Symfony\Component\Cache\Adapter\RedisAdapter;
-use Symfony\Component\Cache\Adapter\RedisTagAwareAdapter;
-use Symfony\Component\Cache\Adapter\TagAwareAdapter;
-use Symfony\Component\Cache\Adapter\TagAwareAdapterInterface;
 use Symfony\Component\Cache\Exception\CacheException;
-use Symfony\Component\Cache\Exception\InvalidArgumentException;
 
 /**
- * The class that apply a bounce.
+ * The class that apply a bounce in standalone mode.
  *
  * @author    CrowdSec team
  *
@@ -27,9 +19,6 @@ use Symfony\Component\Cache\Exception\InvalidArgumentException;
  */
 class StandaloneBounce extends AbstractBounce
 {
-    /** @var TagAwareAdapterInterface|null */
-    protected $cacheAdapter;
-
     /**
      * Initialize the bouncer.
      *
@@ -56,67 +45,9 @@ class StandaloneBounce extends AbstractBounce
             }
             $configs['trust_ip_forward_array'] = $finalForwardConfigs;
         }
-        $this->settings = $configs;
-
-        $this->settings = array_merge($this->settings, $forcedConfigs);
-        $this->setDebug($this->getBoolSettings('debug_mode'));
-        $this->setDisplayErrors($this->getBoolSettings('display_errors'));
-        $this->initLogger();
+        $this->settings = array_merge($configs, $forcedConfigs);
 
         return $this->getBouncerInstance($this->settings);
-    }
-
-    /**
-     * @throws CacheException
-     * @throws ErrorException|BouncerException
-     */
-    private function getCacheAdapterInstance(bool $forceReload = false): TagAwareAdapterInterface
-    {
-        // Singleton for this function
-        if ($this->cacheAdapter && !$forceReload) {
-            return $this->cacheAdapter;
-        }
-
-        $cacheSystem = $this->getStringSettings('cache_system');
-        switch ($cacheSystem) {
-            case Constants::CACHE_SYSTEM_PHPFS:
-                $this->cacheAdapter = new TagAwareAdapter(
-                    new PhpFilesAdapter('', 0, $this->getStringSettings('fs_cache_path'))
-                );
-                break;
-
-            case Constants::CACHE_SYSTEM_MEMCACHED:
-                $memcachedDsn = $this->getStringSettings('memcached_dsn');
-                if (empty($memcachedDsn)) {
-                    throw new BouncerException('The selected cache technology is Memcached.' .
-                                               ' Please set a Memcached DSN or select another cache technology.');
-                }
-
-                $this->cacheAdapter = new MemcachedTagAwareAdapter(
-                    new MemcachedAdapter(MemcachedAdapter::createConnection($memcachedDsn))
-                );
-                break;
-
-            case Constants::CACHE_SYSTEM_REDIS:
-                $redisDsn = $this->getStringSettings('redis_dsn');
-                if (empty($redisDsn)) {
-                    throw new BouncerException('The selected cache technology is Redis.' .
-                                               ' Please set a Redis DSN or select another cache technology.');
-                }
-
-                try {
-                    $this->cacheAdapter = new RedisTagAwareAdapter((RedisAdapter::createConnection($redisDsn)));
-                } catch (InvalidArgumentException $e) {
-                    throw new BouncerException('Error when connecting to Redis.' .
-                                               ' Please fix the Redis DSN or select another cache technology.');
-                }
-                break;
-
-            default:
-                throw new BouncerException("Unknown selected cache technology: $cacheSystem");
-        }
-
-        return $this->cacheAdapter;
     }
 
     /**
@@ -133,67 +64,20 @@ class StandaloneBounce extends AbstractBounce
             return $this->bouncer;
         }
         $this->settings = array_merge($this->settings, $settings);
-        $bouncingLevel = $this->getStringSettings('bouncing_level');
         $apiUserAgent = 'Standalone CrowdSec PHP Bouncer/' . Constants::VERSION;
-        $apiTimeout = $this->getIntegerSettings('api_timeout');
 
-        // Init Bouncer instance
-        switch ($bouncingLevel) {
-            case Constants::BOUNCING_LEVEL_DISABLED:
-                $maxRemediationLevel = Constants::REMEDIATION_BYPASS;
-                break;
-            case Constants::BOUNCING_LEVEL_FLEX:
-                $maxRemediationLevel = Constants::REMEDIATION_CAPTCHA;
-                break;
-            case Constants::BOUNCING_LEVEL_NORMAL:
-                $maxRemediationLevel = Constants::REMEDIATION_BAN;
-                break;
-            default:
-                throw new BouncerException("Unknown $bouncingLevel");
-        }
-        // Instantiate the cache system
-        $this->cacheAdapter = $this->getCacheAdapterInstance($forceReload);
+        $this->settings['api_user_agent'] = $apiUserAgent;
+        $bouncerConfigs = $this->prepareBouncerConfigs();
+
         // Instantiate bouncer
-        $this->bouncer = new Bouncer($this->cacheAdapter, $this->logger, null, $this->settings);
-        // Validate settings
-        $this->bouncer->configure([
-            // LAPI connection
-            'api_key' => $this->getStringSettings('api_key'),
-            'api_url' => $this->getStringSettings('api_url'),
-            'api_user_agent' => $apiUserAgent,
-            'api_timeout' => $apiTimeout > 0 ? $apiTimeout : Constants::API_TIMEOUT,
-            'use_curl' => $this->getBoolSettings('use_curl'),
-            // Debug
-            'debug_mode' => $this->getBoolSettings('debug_mode'),
-            'log_directory_path' => $this->getStringSettings('log_directory_path'),
-            'forced_test_ip' => $this->getStringSettings('forced_test_ip'),
-            'forced_test_forwarded_ip' => $this->getStringSettings('forced_test_forwarded_ip'),
-            'display_errors' => $this->getBoolSettings('display_errors'),
-            // Bouncer
-            'bouncing_level' => $bouncingLevel,
-            'trust_ip_forward_array' => $this->getArraySettings('trust_ip_forward_array'),
-            'fallback_remediation' => $this->getStringSettings('fallback_remediation'),
-            'max_remediation_level' => $maxRemediationLevel,
-            // Cache settings
-            'stream_mode' => $this->getBoolSettings('stream_mode'),
-            'cache_system' => $this->getStringSettings('cache_system'),
-            'fs_cache_path' => $this->getStringSettings('fs_cache_path'),
-            'redis_dsn' => $this->getStringSettings('redis_dsn'),
-            'memcached_dsn' => $this->getStringSettings('memcached_dsn'),
-            'clean_ip_cache_duration' => $this->getIntegerSettings('clean_ip_cache_duration'),
-            'bad_ip_cache_duration' => $this->getIntegerSettings('bad_ip_cache_duration'),
-            'captcha_cache_duration' => $this->getIntegerSettings('captcha_cache_duration'),
-            'geolocation_cache_duration' => $this->getIntegerSettings('geolocation_cache_duration'),
-            // Geolocation
-            'geolocation' => $this->getArraySettings('geolocation'),
-        ]);
+        $this->bouncer = new Bouncer($bouncerConfigs, $this->logger);
 
         return $this->bouncer;
     }
 
-    public function initLogger(): void
+    public function initLogger(array $configs): void
     {
-        $this->initLoggerHelper($this->getStringSettings('log_directory_path'), 'php_standalone_bouncer');
+        $this->initLoggerHelper($configs, 'php_standalone_bouncer');
     }
 
     /**
@@ -474,18 +358,24 @@ class StandaloneBounce extends AbstractBounce
             throw new BouncerException("$errstr (Error level: $errno)");
         });
         try {
-            $this->init($configs);
-            $this->run();
-            $result = true;
+            $this->settings = $configs;
+            $this->initLogger($configs);
+            if ($this->shouldBounceCurrentIp()) {
+                $this->init($configs);
+                $this->run();
+                $result = true;
+            }
         } catch (Exception $e) {
-            $this->logger->error('', [
-                'type' => 'EXCEPTION_WHILE_BOUNCING',
-                'message' => $e->getMessage(),
-                'code' => $e->getCode(),
-                'file' => $e->getFile(),
-                'line' => $e->getLine(),
-            ]);
-            if ($this->displayErrors) {
+            if ($this->logger) {
+                $this->logger->error('', [
+                    'type' => 'EXCEPTION_WHILE_BOUNCING',
+                    'message' => $e->getMessage(),
+                    'code' => $e->getCode(),
+                    'file' => $e->getFile(),
+                    'line' => $e->getLine(),
+                ]);
+            }
+            if (!empty($configs['display_errors'])) {
                 throw $e;
             }
         }