+++ tests/Integration/IpVerificationTest.php
@@ -4,10 +4,8 @@ declare(strict_types=1);
 
 namespace CrowdSecBouncer\Tests\Integration;
 
-use CrowdSecBouncer\ApiCache;
-use CrowdSecBouncer\ApiClient;
 use CrowdSecBouncer\Bouncer;
-use PHPUnit\Framework\MockObject\MockObject;
+use CrowdSecBouncer\Constants;
 use PHPUnit\Framework\TestCase;
 use Psr\Log\LoggerInterface;
 
@@ -19,46 +17,47 @@ final class IpVerificationTest extends TestCase
     /** @var LoggerInterface */
     private $logger;
 
-    private $useCurl = false;
+    /** @var bool  */
+    private $useCurl;
 
     protected function setUp(): void
     {
         $this->logger = TestHelpers::createLogger();
-        $this->useCurl = getenv('USE_CURL');
-        $this->watcherClient = new WatcherClient($this->logger, ['use_curl' => $this->useCurl]);
-        $this->watcherClient->configure();
+        $this->useCurl = (bool) getenv('USE_CURL');
+        $this->watcherClient = new WatcherClient(['use_curl' => $this->useCurl], $this->logger);
     }
 
-    public function cacheAdapterProvider(): array
+    public function cacheAdapterConfigProvider(): array
     {
-        return TestHelpers::cacheAdapterProvider();
+        return TestHelpers::cacheAdapterConfigProvider();
     }
 
     /**
      * @group integration
-     * @dataProvider cacheAdapterProvider
+     * @dataProvider cacheAdapterConfigProvider
      */
-    public function testCanVerifyIpInLiveModeWithCacheSystem($cacheAdapter, $origCacheName): void
+    public function testCanVerifyIpInLiveModeWithCacheSystem($cacheAdapterName, $origCacheName): void
     {
         // Init context
-
         $this->watcherClient->setInitialState();
-        $cacheAdapter->clear();
 
         // Init bouncer
-
-        /** @var ApiClient */
-        $apiClientMock = $this->getMockBuilder(ApiClient::class)
-            ->setConstructorArgs([$this->logger, ['use_curl' => $this->useCurl]])
-            ->enableProxyingToOriginalMethods()
-            ->getMock();
-        $apiCache = new ApiCache($this->logger, $apiClientMock, $cacheAdapter);
-        $bouncerConfig = [
+        $bouncerConfigs = [
             'api_key' => TestHelpers::getBouncerKey(),
             'api_url' => TestHelpers::getLapiUrl(),
+            'use_curl' => $this->useCurl,
+            'api_user_agent' => TestHelpers::UNIT_TEST_AGENT_PREFIX . '/' . Constants::BASE_USER_AGENT,
+            'cache_system' => $cacheAdapterName,
+            'redis_dsn' => getenv('REDIS_DSN'),
+            'memcached_dsn' =>  getenv('MEMCACHED_DSN'),
+            'fs_cache_path' => TestHelpers::PHP_FILES_CACHE_ADAPTER_DIR
         ];
-        $bouncer = new Bouncer(null, $this->logger, $apiCache);
-        $bouncer->configure($bouncerConfig);
+
+        $bouncer = new Bouncer($bouncerConfigs, $this->logger);
+
+        // Test cache adapter
+        $cacheAdapter = $bouncer->getCacheAdapter();
+        $cacheAdapter->clear();
 
         switch ($origCacheName) {
             case 'PhpFilesAdapter':
@@ -85,10 +84,6 @@ final class IpVerificationTest extends TestCase
             default:
                 break;
         }
-
-        // At the end of test, we should have exactly 3 "cache miss")
-        /** @var MockObject $apiClientMock */
-        $apiClientMock->expects($this->exactly(4))->method('getFilteredDecisions');
 
         $this->assertEquals(
             'ban',
@@ -127,13 +122,13 @@ final class IpVerificationTest extends TestCase
         $this->assertEquals('ban', $remediation3rdCall);
 
         // Reconfigure the bouncer to set maximum remediation level to "captcha"
-        $bouncerConfig['max_remediation_level'] = 'captcha';
-        $bouncer->configure($bouncerConfig);
+        $bouncerConfigs['max_remediation_level'] = 'captcha';
+        $bouncer = new Bouncer( $bouncerConfigs, $this->logger);
         $cappedRemediation = $bouncer->getRemediationForIp(TestHelpers::BAD_IP);
         $this->assertEquals('captcha', $cappedRemediation, 'The remediation for the banned IP should now be "captcha"');
         // Reset the max remediation level to its origin state
-        unset($bouncerConfig['max_remediation_level']);
-        $bouncer->configure($bouncerConfig);
+        unset($bouncerConfigs['max_remediation_level']);
+        $bouncer = new Bouncer($bouncerConfigs, $this->logger);
 
         $this->logger->info('', ['message' => 'set "Large IPV4 range banned" state']);
         $this->watcherClient->deleteAllDecisions();
@@ -151,31 +146,29 @@ final class IpVerificationTest extends TestCase
 
     /**
      * @group integration
-     * @dataProvider cacheAdapterProvider
+     * @dataProvider cacheAdapterConfigProvider
      */
-    public function testCanVerifyIpInStreamModeWithCacheSystem($cacheAdapter, $origCacheName): void
+    public function testCanVerifyIpInStreamModeWithCacheSystem($cacheAdapterName, $origCacheName): void
     {
         // Init context
-
         $this->watcherClient->setInitialState();
-        $cacheAdapter->clear();
-
         // Init bouncer
-
-        /** @var ApiClient */
-        $apiClientMock = $this->getMockBuilder(ApiClient::class)
-            ->setConstructorArgs([$this->logger, ['use_curl' => $this->useCurl]])
-            ->enableProxyingToOriginalMethods()
-            ->getMock();
-        $apiCache = new ApiCache($this->logger, $apiClientMock, $cacheAdapter);
-
-        $bouncerConfig = [
+        $bouncerConfigs = [
             'api_key' => TestHelpers::getBouncerKey(),
             'api_url' => TestHelpers::getLapiUrl(),
+            'api_user_agent' => TestHelpers::UNIT_TEST_AGENT_PREFIX . '/' . Constants::BASE_USER_AGENT,
             'stream_mode' => true,
+            'use_curl' => $this->useCurl,
+            'cache_system' => $cacheAdapterName,
+            'redis_dsn' => getenv('REDIS_DSN'),
+            'memcached_dsn' =>  getenv('MEMCACHED_DSN'),
+            'fs_cache_path' => TestHelpers::PHP_FILES_CACHE_ADAPTER_DIR
         ];
-        $bouncer = new Bouncer(null, $this->logger, $apiCache);
-        $bouncer->configure($bouncerConfig);
+
+        $bouncer = new Bouncer($bouncerConfigs, $this->logger);
+        // Test cache adapter
+        $cacheAdapter = $bouncer->getCacheAdapter();
+        $cacheAdapter->clear();
 
         switch ($origCacheName) {
             case 'PhpFilesAdapter':
@@ -203,10 +196,6 @@ final class IpVerificationTest extends TestCase
                 break;
         }
         // As we are in stream mode, no live call should be done to the API.
-
-        /** @var MockObject $apiClientMock */
-        $apiClientMock->expects($this->exactly(0))->method('getFilteredDecisions');
-
         // Warm BlockList cache up
 
         $bouncer->refreshBlocklistCache();
@@ -221,13 +210,12 @@ final class IpVerificationTest extends TestCase
         );
 
         // Reconfigure the bouncer to set maximum remediation level to "captcha"
-        $bouncerConfig['max_remediation_level'] = 'captcha';
-        $bouncer->configure($bouncerConfig);
+        $bouncerConfigs['max_remediation_level'] = 'captcha';
+        $bouncer = new Bouncer($bouncerConfigs, $this->logger);
         $cappedRemediation = $bouncer->getRemediationForIp(TestHelpers::BAD_IP);
         $this->assertEquals('captcha', $cappedRemediation, 'The remediation for the banned IP should now be "captcha"');
-        unset($bouncerConfig['max_remediation_level']);
-        $bouncer->configure($bouncerConfig);
-
+        unset($bouncerConfigs['max_remediation_level']);
+        $bouncer = new Bouncer($bouncerConfigs, $this->logger);
         $this->assertEquals(
             'bypass',
             $bouncer->getRemediationForIp(TestHelpers::CLEAN_IP),
@@ -250,8 +238,6 @@ final class IpVerificationTest extends TestCase
         $this->logger->debug('', ['message' => 'Refresh 2nd time the cache. Nothing should append.']);
         $bouncer->refreshBlocklistCache();
 
-        $this->logger->debug('', ['message' => 'Refresh 3rd time the cache. Nothing should append.']);
-        $bouncer->refreshBlocklistCache();
 
         $this->assertEquals(
             'ban',
@@ -266,26 +252,19 @@ final class IpVerificationTest extends TestCase
         );
 
         // Setup an new instance.
-
-        /** @var ApiClient */
-        $apiClientMock2 = $this->getMockBuilder(ApiClient::class)
-            ->setConstructorArgs([$this->logger,['use_curl' => $this->useCurl]])
-            ->enableProxyingToOriginalMethods()
-            ->getMock();
-        $apiCache2 = new ApiCache($this->logger, $apiClientMock2, $cacheAdapter);
-
-        $bouncerConfig = [
+        $bouncerConfigs = [
             'api_key' => TestHelpers::getBouncerKey(),
             'api_url' => TestHelpers::getLapiUrl(),
             'stream_mode' => true,
+            'use_curl' => $this->useCurl,
+            'api_user_agent' => TestHelpers::UNIT_TEST_AGENT_PREFIX . '/' . Constants::BASE_USER_AGENT,
+            'cache_system' => $cacheAdapterName,
+            'redis_dsn' => getenv('REDIS_DSN'),
+            'memcached_dsn' =>  getenv('MEMCACHED_DSN'),
+            'fs_cache_path' => TestHelpers::PHP_FILES_CACHE_ADAPTER_DIR
         ];
-        $bouncer = new Bouncer(null, $this->logger, $apiCache2);
-        $bouncer->configure($bouncerConfig);
 
-        // The cache should still be warmed up, even for a new instance
-
-        /** @var MockObject $apiClientMock2 */
-        $apiClientMock2->expects($this->exactly(0))->method('getFilteredDecisions');
+        $bouncer = new Bouncer($bouncerConfigs, $this->logger);
 
         $this->assertEquals(
             'ban',